refactor: 클래스 패키지로 분리
- request 패키지에는 FileInfo, HttpRequest 클래스.
- response 패키지에는 ContentType, HttpResponseBody, HttpResponseHeader
  클래스.
