+++ src/main/java/request/FileInfo.java
@@ -1,4 +1,4 @@
-package webserver;
+package request;
 
 import java.io.File;
