+++ src/test/java/webserver/HttpRequestTest.java
@@ -3,6 +3,8 @@ package webserver;
 import db.Database;
 import model.User;
 import org.junit.jupiter.api.*;
+import request.HttpRequest;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class HttpRequestTest {
